requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
faiss-cpu
pypdf


